+++ source/dea-ui/ui/src/components/buttons/DownloadButton.tsx
@@ -5,12 +5,14 @@
 
 import { DownloadDTO } from '@aws/dea-app/lib/models/case-file';
 import { CaseStatus } from '@aws/dea-app/lib/models/case-status';
-import { Button, Spinner } from '@cloudscape-design/components';
+import { Button, SpaceBetween, Spinner } from '@cloudscape-design/components';
+import { useState } from 'react';
 import { useAvailableEndpoints } from '../../api/auth';
 import { getPresignedUrl, useGetCaseActions } from '../../api/cases';
 import { commonLabels, fileOperationsLabels } from '../../common/labels';
 import { useNotifications } from '../../context/NotificationsContext';
 import { canDownloadFiles, canRestoreFiles } from '../../helpers/userActionSupport';
+import { FormFieldModal } from '../common-components/FormFieldModal';
 
 export interface DownloadButtonProps {
   readonly caseId: string;
@@ -27,14 +29,24 @@ function DownloadButton(props: DownloadButtonProps): JSX.Element {
   const { pushNotification } = useNotifications();
   const userActions = useGetCaseActions(props.caseId);
   const availableEndpoints = useAvailableEndpoints();
+  const [downloadReasonModalOpen, setDownloadReasonModalOpen] = useState(false);
+  const [downloadReason, setDownloadReason] = useState('');
 
   async function downloadFilesHandler() {
     const sleep = (delay: number) => new Promise((resolve) => setTimeout(resolve, delay));
     try {
+      setDownloadReasonModalOpen(false);
       props.downloadInProgressCallback(true);
+
+      let allFilesDownloaded = true;
       for (const file of props.selectedFiles) {
         try {
-          const downloadResponse = await getPresignedUrl({ caseUlid: file.caseUlid, ulid: file.ulid });
+          const downloadResponse = await getPresignedUrl({
+            caseUlid: file.caseUlid,
+            ulid: file.ulid,
+            downloadReason: downloadReason,
+          });
+
           if (!downloadResponse.downloadUrl) {
             if (downloadResponse.isRestoring) {
               pushNotification('info', fileOperationsLabels.restoreInProgress(file.fileName));
@@ -57,27 +69,58 @@ function DownloadButton(props: DownloadButtonProps): JSX.Element {
         } catch (e) {
           pushNotification('error', fileOperationsLabels.downloadFailed(file.fileName));
           console.log(`failed to download ${file.fileName}`, e);
+          allFilesDownloaded = false;
         }
+      }
+
+      if (allFilesDownloaded) {
+        pushNotification('success', fileOperationsLabels.downloadSucceeds(props.selectedFiles.length));
       }
     } finally {
       props.downloadInProgressCallback(false);
+      setDownloadReason('');
       props.selectedFilesCallback([]);
     }
   }
 
   return (
-    <Button
-      data-testid="download-file-button"
-      variant="primary"
-      onClick={downloadFilesHandler}
-      disabled={
-        props.downloadInProgress ||
-        !(canDownloadFiles(userActions?.data?.actions) && props.caseStatus === CaseStatus.ACTIVE)
-      }
-    >
-      {commonLabels.downloadButton}
-      {props.downloadInProgress ? <Spinner size="normal" /> : null}
-    </Button>
+    <SpaceBetween direction="horizontal" size="xs">
+      <FormFieldModal
+        modalTestId="download-file-reason-modal"
+        inputTestId="download-file-reason-modal-input"
+        cancelButtonTestId="download-file-reason-modal-cancel-button"
+        primaryButtonTestId="download-file-reason-modal-primary-button"
+        isOpen={downloadReasonModalOpen}
+        title={fileOperationsLabels.downloadFileReasonLabel}
+        inputHeader={fileOperationsLabels.downloadFileReasonInputHeader}
+        inputDetails={fileOperationsLabels.downloadFileReasonInputDetails}
+        inputField={downloadReason}
+        setInputField={setDownloadReason}
+        confirmAction={downloadFilesHandler}
+        confirmButtonText={commonLabels.downloadButton}
+        cancelAction={() => {
+          // close modal and delete any reason inputted
+          setDownloadReasonModalOpen(false);
+          setDownloadReason('');
+        }}
+        cancelButtonText={commonLabels.cancelButton}
+      />
+      <Button
+        data-testid="download-file-button"
+        variant="primary"
+        onClick={() => {
+          setDownloadReasonModalOpen(true);
+        }}
+        disabled={
+          props.selectedFiles.length === 0 ||
+          props.downloadInProgress ||
+          !(canDownloadFiles(userActions?.data?.actions) && props.caseStatus === CaseStatus.ACTIVE)
+        }
+      >
+        {commonLabels.downloadButton}
+        {props.downloadInProgress ? <Spinner size="normal" /> : null}
+      </Button>
+    </SpaceBetween>
   );
 }